WIP: Prototype of a "relay" component
This change prototypes a relay component. The goal of this component
is to be a drop-in replacement for jaeger-collector, but instead of
writing to cassandra (or another writer) relay the spans to another
jaeger-collector (or compatible) instance. The relay offers a choice
of thrift-tChannel or thrift-HTTP sender, which can be configured at
startup.

TODOS:
TODO: The relay is its own component. It should be merged into the
jaeger-collector repo, by generalizing the concept of writer/processor.

TODO: Rather than hard-coding the available senders using switch-case
statements in main, and type-specific methods in builder, use the
factory pattern similar to how the storage factory is implemented
in collector.