Add a bounded in-memory queue between receivers and sender
This change adds a bounded in-memory queue to hold span batches
until the sender is able to send them out. There are several
configuration options to control behavior, including queue
size, number of workers, and a retry policy in case of failure
to send (retry or don't retry and drop).

The spanBatchProcessor is adapted from the spanProcessor in the
collector app, however the main difference is that while the
former operates upon span batches, the latter operates upon
individual spans.

In order to merge relay code base into collector code base, the
interface for span writer would need to be changed to accept
span batches, and possibly write a batch at a time (is there
any benefit to writing one span at a time to cassandra from
50 concurrent threads?)