Add multi-destination support
This change adds multi-destination support to the relay.
Reworked the builder config parsing logic to make it possible
to supply configuration options for multiple processors.
Added a MultiSpanBatchProcessor that offers batches one at
a time to the registered processors. This should not be a
concern, since each processor has a in-memory buffer so
it enqueues and immediately returns. Separated out metrics
on a per-processor basis using the name supplied by user.